+++ backend-rust/CHANGELOG.md
@@ -8,6 +8,8 @@ Database schema version: 18
 
 ### Added
 
+- Add query `SuspendedValidators::suspendedValidators`.
+- Add query `SuspendedValidators::primedForSuspensionValidators`.
 - Add database migration 18 adding table tracking baker pool and passive stake for every reward period.
 - Add query `PassiveDelegation::apy` and `BakerPool::apy`.
 
@@ -17,6 +19,7 @@ Database schema version: 17
 
 ### Added
 
+- Add database migration 17 adding a table tracking reward metrics.
 - Add database migration 16 adding a table tracking commission rates for passive delegation and adding an index to retrieve passive delegators efficiently from the accounts table.
 - Add query `PassiveDelegation::delegators`, `PassiveDelegation::delegatorCount`,
 `PassiveDelegation::commissionRates`, `PassiveDelegation::delegatedStake` and `PassiveDelegation::delegatedStakePercentage`.